+++ src/containers/landing/lunch-menu/school-card/school-card.tsx
